requirements.txt
asgiref==3.6.0
certifi==2022.12.7
charset-normalizer==3.1.0
coreapi==2.3.3
coreschema==0.0.4
cx-Oracle==8.3.0
Django==4.1.7
django-cors-headers==3.14.0
django-db-connection-pool==1.2.2
django-environ==0.10.0
djangorestframework==3.14.0
drf-yasg==1.21.5
greenlet==2.0.2
idna==3.4
inflection==0.5.1
itypes==1.2.0
JayDeBeApi==1.2.3
Jinja2==3.1.2
JPype1==1.4.1
MarkupSafe==2.1.2
mysqlclient==2.1.1
packaging==23.0
psycopg2==2.9.5
PyMySQL==1.0.3
python-dotenv==1.0.0
pytz==2023.3
requests==2.28.2
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.7
SQLAlchemy==2.0.7
sqlparams==5.1.0
sqlparse==0.4.3
tenacity==8.2.2
typing_extensions==4.5.0
tzdata==2023.3
uritemplate==4.1.1
urllib3==1.26.15
